requirements.txt
streamlit
python-dotenv
anthropic
PyPDF2
pdf2image
Pillow
pdfminer.six
PyMuPDF

